Move linalg based tensor factory methods to common.py
